Refactor contract API into reporting, truth auction, and fork modules
- Extracted reporting logic (deposits, escrow ops, and fork proof helpers) from `ui/ts/contracts.ts` into `ui/ts/contracts/reporting.ts`
- Moved truth-auction query helpers into `ui/ts/contracts/truthAuctions.ts`
- Moved fork-auction action execution helper into `ui/ts/contracts/securityPoolActions.ts`
- Added shared decoders/type helpers in `ui/ts/contracts/decoders.ts` and updated imports/exports in `ui/ts/contracts.ts`
- Updated `ui/ts/tests/contracts.test.ts` to match the reorganized contract module structure
